+++ onmt/data/SampledDataset.lua
@@ -3,15 +3,29 @@
 local SampledDataset = torch.class("SampledDataset")
 
 local options = {
-  {'-sample',              0, [[Number of instances to sample from train data in each epoch]]},
-  {'-sample_w_ppl',    false, [[use ppl as probability distribution when sampling]]},
-  {'-sample_w_ppl_init',  15, [[start perplexity-based sampling when average train perplexity per batch falls below this value]]},
-  {'-sample_w_ppl_max', -1.5, [[when greater than 0, instances with perplexity above this value will be considered as noise and ignored;
-                                when less than 0, mode + (-sample_w_ppl_max) * stdev will be used as threshold]]},
+  {
+    '-sample', 0,
+    [[Number of instances to sample from train data in each epoch.]]
+  },
+  {
+    '-sample_w_ppl', false,
+    [[If set, ese perplexity as a probability distribution when sampling.]]
+  },
+  {
+    '-sample_w_ppl_init', 15,
+    [[Start perplexity-based sampling when average train perplexity per batch
+      falls below this value.]]
+  },
+  {
+    '-sample_w_ppl_max', -1.5,
+    [[When greater than 0, instances with perplexity above this value will be
+      considered as noise and ignored; when less than 0, mode + (-sample_w_ppl_max) * stdev
+      will be used as threshold.]]
+  }
 }
 
 function SampledDataset.declareOpts(cmd)
-  cmd:setCmdLineOptions(options, 'SampledDataset')
+  cmd:setCmdLineOptions(options, 'Sampled dataset')
 end
 
 --[[ Initialize a data object given aligned tables of IntTensors `srcData`
@@ -164,15 +178,20 @@ function SampledDataset:sample()
   _G.logger:info('Sampled ' .. self.sampled:size(1) .. ' instances')
 
   -- Prepares batches in terms of range within self.src and self.tgt.
+  local batchesCapacity = 0
+  local batchesOccupation = 0
   self.batchRange = {}
   local offset = 0
   local sampleCntBegin = 1
   local batchSize = 1
-  local sourceLength = -1
+  local maxSourceLength = -1
   for i = 1, #self.src do
     for j = 1, self.sampledCnt[i] do
-      if batchSize == self.maxBatchSize or self.src[i]:size(1) ~= sourceLength then
+      local sourceLength = self.src[i]:size(1)
+      if batchSize == self.maxBatchSize or offset == 1 or
+         (not(self.uneven_batches) and self.src[i]:size(1) ~= maxSourceLength) then
         if offset > 0 then
+          batchesCapacity = batchesCapacity + batchSize * maxSourceLength
           local batchEnd = (j == 1) and i - 1 or i
           local sampleCntEnd = (j == 1) and self.sampledCnt[i - 1] or j - 1
           table.insert(self.batchRange, {
@@ -185,14 +204,17 @@ function SampledDataset:sample()
         end
         offset = i
         batchSize = 1
-        sourceLength = self.src[i]:size(1)
+        maxSourceLength = -1
       else
         batchSize = batchSize + 1
       end
+      batchesOccupation = batchesOccupation + sourceLength
+      maxSourceLength = math.max(maxSourceLength, sourceLength)
     end
   end
   -- Catch last batch.
   if offset < #self.src then
+    batchesCapacity = batchesCapacity + batchSize * maxSourceLength
     table.insert(self.batchRange, {
       ["begin"] = offset,
       ["end"] = #self.src,
@@ -202,6 +224,7 @@ function SampledDataset:sample()
   end
 
   _G.logger:info('Prepared ' .. #self.batchRange .. ' batches')
+  return #self.batchRange, batchesOccupation / batchesCapacity
 end
 
 --[[ Get perplexity. ]]
@@ -244,7 +267,7 @@ function SampledDataset:setBatchSize(maxBatchSize, uneven_batches)
     end
   end
 
-  self:sample()
+  return self:sample()
 end
 
 --[[ Return number of sampled instances. ]]