+++ tools/utils/separators.lua
@@ -2,5 +2,7 @@ return {
   joiner_marker = '￭',
   joiner_marker_substitute = '■',
   feat_marker = '￨',
-  feat_marker_substitute = '│'
+  feat_marker_substitute = '│',
+  BOT = "<w>",
+  EOT = "</w>"
 }