+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Allow to change the state of word embeddings optimization during a retraining
 * Check consistency of option settings when training from existing model
 * Save and Restore random number generator states in checkpoints
+* Improve duplicate messages in parallel mode
 
 ## [v0.5.3](https://github.com/OpenNMT/OpenNMT/releases/tag/v0.5.3) (2017-03-30)